requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
pytest==7.4.3
pytest-django==4.7.0 